Added polymorphic mediable wip

+++ app/Jobs/CheckPicture.php
@@ -7,6 +7,7 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
+use Illuminate\Support\Facades\Auth;
 use Sightengine\SightengineClient;
 
 class CheckPicture implements ShouldQueue
@@ -37,22 +38,28 @@ class CheckPicture implements ShouldQueue
     {
         //analyze the locally stored image for nudity
         // wad : weapon, alcool, drugs
-        if (config('services.sightengine.key') !== '') {
-            $sightEngine = new SightengineClient(config('services.sightengine.user'),
-                config('services.sightengine.key'));
-            $path = $this->media->getPath();
-            $imageCheck = $sightEngine->check(['nudity', 'wad', 'offensive', 'face-attributes'])
-                ->set_file($path);
+        if (config('app.env') == 'production') {
+            if (config('services.sightengine.key') !== '') {
+                $sightEngine = new SightengineClient(
+                    config('services.sightengine.user'),
+                    config('services.sightengine.key')
+                );
+                $path = $this->media->getPath();
+                $imageCheck = $sightEngine->check(['nudity', 'wad', 'offensive', 'face-attributes'])
+                    ->set_file($path);
 
-            if ($imageCheck->status !== 'success') {
-                // delete picture
-                $mediaItems = $this->recipe->getMedia();
-                $media = $mediaItems->where('id', $this->media->id)->first();
-                $media->delete();
-            } else {
-                $mediaItems = $this->recipe->getMedia();
-                $media = $mediaItems->where('id', $this->media->id)->first();
-                $media->setCustomProperty('checked', true);
+                if ($imageCheck->status !== 'success') {
+                    // delete picture
+                    $mediaItems = $this->recipe->getMedia();
+                    $media = $mediaItems->where('id', $this->media->id)->first();
+                    $this->recipe->medias()->detach([$media->id]);
+                    Auth::user()->medias()->detach([$media->id]);
+                    $media->delete();
+                } else {
+                    $mediaItems = $this->recipe->getMedia();
+                    $media = $mediaItems->where('id', $this->media->id)->first();
+                    $media->setCustomProperty('checked', true);
+                }
             }
         }
     }
